Restructured processing file system, fixed tests
