+++ src/data_processing/preprocess_data.py
@@ -1,8 +1,6 @@
-import os
 import logging
 import argparse
 import pandas as pd
-import glob
 from sklearn.model_selection import train_test_split
 
 from utils.auxilliary_preprocessing import *
