+++ src/data_processing/utils/text_normalization.py
@@ -182,7 +182,7 @@ def normalize_text_columns(df: pd.DataFrame, text_columns: list[str], language_c
         expanded_columns = text_columns
 
     for col in expanded_columns:
-        new_col = f"{col}_normalized"
+        new_col = f"{col[:-3]}_normalized_{col[-2:]}"  # create a new column for the normalized text
         df[new_col] = df[col]  # create a new column for the normalized text
         logger.info(f"Normalizing column {col} into new column {new_col}")
         section_header = language_config[col[-2:]].get('section_header_to_remove')
@@ -197,3 +197,44 @@ def normalize_text_columns(df: pd.DataFrame, text_columns: list[str], language_c
         df = apply_fn_and_count_affected_rows(df, new_col, remove_extra_whitespace)
     
     return df
+
+
+def preprocess_records_into_text(df, languages=['en', 'fr'], keyword_prefixes=['keywords:', 'mots-clés associés:'], output_format='sequential'):
+    '''
+    Combines text columns from the DataFrame to create a text representation of the records in the specified languages.
+
+    Parameters:
+    - df: pandas DataFrame containing the records.
+    - languages: List of languages to include in the text representation. Supported values are 'en' for English and 'fr' for French. Default is [en, fr].
+    - keyword_prefixes: List of prefixes to use for the keywords in each language. The length of this list should match the number of languages specified. Default is ['keywords:', 'mots-clés associés:'] for English and French respectively.
+    - output_format: Method to combine the text columns. Supported values are 'sequential' for combining in the order of languages and 'parallel' for combining all languages together. Default is 'sequential'.
+
+    Returns:
+    - pandas DataFrame: DataFrame containing the text representations of the records.
+    '''
+    # confirm that the length of keyword_prefixes matches the number of languages
+    if len(keyword_prefixes) != len(languages):
+        raise ValueError("Length of keyword_prefixes must match the number of languages specified.")
+
+    selected_columns = []
+    for lang in languages:
+        selected_columns.extend([f'features_properties_title_{lang}', f'features_properties_description_normalized_{lang}', f'features_properties_keywords_{lang}'])
+    
+    # send warning if any of the selected columns are not in the DataFrame
+    missing_columns = [col for col in selected_columns if col not in df.columns]
+    if missing_columns:
+        logger.error(f"Columns {missing_columns} used for text processing are not found in the DataFrame.")
+        return pd.Series(["\n".join(["\n\n\n" for _ in languages])] * len(df))
+
+    df = df[selected_columns]
+    if output_format == 'sequential':
+        return df.apply(lambda x: "\n".join(
+            [f"{x[f'features_properties_title_{lang}']}\n{keyword_prefixes[i]}{x[f'features_properties_keywords_{lang}'] if isinstance(x[f'features_properties_keywords_{lang}'], str) else ', '.join(x[f'features_properties_keywords_{lang}']) if isinstance(x[f'features_properties_keywords_{lang}'], list) else ''}\n{x[f'features_properties_description_normalized_{lang}']}" for i, lang in enumerate(languages)]), axis=1)
+    elif output_format == 'parallel':
+        return df.apply(lambda x: "\n".join( 
+            [f"{x[f'features_properties_title_{lang}']}" for lang in languages] + 
+            [f"{keyword_prefixes[i]}{x[f'features_properties_keywords_{lang}'] if isinstance(x[f'features_properties_keywords_{lang}'], str) else ', '.join(x[f'features_properties_keywords_{lang}']) if isinstance(x[f'features_properties_keywords_{lang}'], list) else ''}" for i, lang in enumerate(languages)] +
+            [f"{x[f'features_properties_description_normalized_{lang}']}" for lang in languages]),
+            axis=1)
+    else:
+        raise ValueError("Invalid processing method. Supported values are 'sequential' and 'parallel'.")
